Move bluez-specific code into bluez crate
