+++ src/api/host.rs
@@ -1,10 +1,20 @@
-use adapter::EventHandler;
 use api::BDAddr;
 use api::peripheral::Peripheral;
 use ::Result;
 
+#[derive(Debug, Copy, Clone)]
+pub enum Event {
+    DeviceDiscovered(BDAddr),
+    DeviceLost(BDAddr),
+    DeviceUpdated(BDAddr),
+    DeviceConnected(BDAddr),
+    DeviceDisconnected(BDAddr),
+}
+
+pub type EventHandler = Box<Fn(Event) + Send>;
+
 pub trait Host {
-    fn watch(&self, handler: EventHandler);
+    fn on_event(&self, handler: EventHandler);
 
     fn start_scan(&self) -> Result<()>;
     fn stop_scan(&self) -> Result<()>;
